+++ src/irmin-test/store.ml
@@ -1002,7 +1002,8 @@ module Make (S : S) = struct
         | `Contents -> Fmt.string ppf "contents"
         | `Node `Hash -> Fmt.string ppf "hash"
         | `Node `Map -> Fmt.string ppf "map"
-        | `Node `Value -> Fmt.string ppf "value")
+        | `Node `Value -> Fmt.string ppf "value"
+        | `Node `Pruned -> Fmt.string ppf "pruned")
       ( = )
 
   let test_tree_caches x () =